requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
autopep8==1.5.5
cachetools==4.2.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-core==1.7.2
googleapis-common-protos==1.56.0
googletrans==3.0.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.44.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
packaging==21.3
promise==2.3
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==1.7.0
pyparsing==3.0.8
python-environ==0.4.54
pytz==2021.1
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.9
vaderSentiment==3.3.2
whitenoise==5.2.0
